Make forms consistent

+++ packages/interfaces/src/common/mongo-collection.ts
@@ -1,17 +1,19 @@
 /* * */
 
-import { AggregationPipeline } from '@/common/aggregation-pipeline.js';
+import { type AggregationPipeline } from '@/common/aggregation-pipeline.js';
 import { HttpException, HttpStatus } from '@tmlmobilidade/consts';
 import { Dates } from '@tmlmobilidade/dates';
 import { MongoConnector } from '@tmlmobilidade/mongo';
 import { generateRandomString } from '@tmlmobilidade/strings';
 import { type UnixTimestamp } from '@tmlmobilidade/types';
-import { AggregateOptions, AggregationCursor, Collection, DeleteOptions, DeleteResult, Document, Filter, FindOptions, IndexDescription, InsertManyResult, InsertOneOptions, InsertOneResult, MongoClientOptions, OptionalUnlessRequiredId, UpdateOptions, UpdateResult, WithId } from 'mongodb';
+import { type AggregateOptions, type AggregationCursor, type Collection, type DeleteOptions, type DeleteResult, type Document, type Filter, type FindOptions, type IndexDescription, type InsertManyResult, type InsertOneOptions, type InsertOneResult, type MongoClientOptions, type OptionalUnlessRequiredId, type UpdateOptions, type UpdateResult, type WithId } from 'mongodb';
 import { z } from 'zod';
 
 /* * */
 
 export abstract class MongoCollectionClass<T extends Document, TCreate, TUpdate> {
+	//
+
 	protected createSchema: null | z.ZodSchema = null;
 	protected mongoCollection: Collection<T>;
 	protected mongoConnector: MongoConnector;
@@ -19,20 +21,19 @@ export abstract class MongoCollectionClass<T extends Document, TCreate, TUpdate>
 
 	/**
 	 * Aggregates documents in the collection.
-	 * @param pipeline - The aggregation pipeline to execute
-	 * @param options - The options for the aggregation operation
-	 * @returns A promise that resolves to an array of aggregated documents
+	 * @param pipeline The aggregation pipeline to execute.
+	 * @param options The options for the aggregation operation.
+	 * @returns A promise that resolves to an array of aggregated documents.
 	 */
 	public async aggregate(pipeline: AggregationPipeline<T>, options?: AggregateOptions & { returnResult?: true }): Promise<T[]>;
 	public async aggregate(pipeline: AggregationPipeline<T>, options: AggregateOptions & { returnResult: false }): Promise<AggregationCursor<T>>;
 	public async aggregate(pipeline: AggregationPipeline<T>, options?: AggregateOptions & { returnResult?: boolean }): Promise<AggregationCursor<T> | T[]> {
-		const aggregation = this.mongoCollection.aggregate(pipeline, options);
-
-		if (options?.returnResult === false) {
-			return aggregation as AggregationCursor<T>;
-		}
-
-		return aggregation.toArray() as Promise<T[]>;
+		// Perform the aggregation pipeline
+		const aggregationResult = this.mongoCollection.aggregate(pipeline, options);
+		// If returnResult is false, return the cursor directly
+		if (options?.returnResult === false) return aggregationResult as AggregationCursor<T>;
+		// Otherwise, return the aggregated documents as an array
+		return aggregationResult.toArray() as Promise<T[]>;
 	}
 
 	/**
@@ -45,18 +46,14 @@ export abstract class MongoCollectionClass<T extends Document, TCreate, TUpdate>
 
 	/**
 	 * Establishes a connection to the Mongo database and initializes the collection.
-	 * @param options Optional Mongo client connection options
-	 * @throws {Error} If connection fails
+	 * @param options Optional Mongo client connection options.
+	 * @throws Error if the environment variable for the database URI is missing or if the connection fails.
 	 */
 	public async connect(options?: MongoClientOptions) {
-		//
-
+		// Extract the database URI from environment variables
 		const dbUri = process.env[this.getEnvName()];
-
-		if (!dbUri) {
-			throw new Error(`Missing ${this.getEnvName()} environment variable`);
-		}
-
+		if (!dbUri) throw new Error(`Missing ${this.getEnvName()} environment variable`);
+		// Attempt to connect to the database
 		try {
 			// Connect to the MongoDB database
 			this.mongoConnector = new MongoConnector(dbUri, options);
@@ -76,8 +73,8 @@ export abstract class MongoCollectionClass<T extends Document, TCreate, TUpdate>
 
 	/**
 	 * Counts documents matching the filter criteria.
-	 * @param filter - The filter criteria to match documents
-	 * @returns A promise that resolves to the count of matching documents
+	 * @param filter The filter criteria to match documents.
+	 * @returns A promise that resolves to the count of matching documents.
 	 */
 	public async count(filter?: Filter<T>): Promise<number> {
 		return await this.mongoCollection.countDocuments(filter);
@@ -85,16 +82,21 @@ export abstract class MongoCollectionClass<T extends Document, TCreate, TUpdate>
 
 	/**
 	 * Deletes a single document by its ID.
-	 * @param id - The ID of the document to delete
-	 * @returns A promise that resolves to the result of the delete operation
+	 * @param id The ID of the document to delete.
+	 * @returns A promise that resolves to the result of the delete operation.
 	 */
-	public async deleteById(id: string, options?: DeleteOptions): Promise<DeleteResult> {
-		const result = await this.deleteOne({ _id: { $eq: id } } as unknown as Filter<T>, options);
-
-		if (!result.acknowledged) {
-			throw new HttpException(HttpStatus.INTERNAL_SERVER_ERROR, 'Failed to delete documents', result);
+	public async deleteById(id: T['_id'], options?: DeleteOptions & { forceIfLocked?: boolean }): Promise<DeleteResult> {
+		// If forceIfLocked is not set then check if the document is locked.
+		// If it is locked, then throw an error to prevent the operation.
+		if (!options?.forceIfLocked) {
+			const isLocked = await this.isLockedById(id);
+			if (isLocked) throw new HttpException(HttpStatus.FORBIDDEN, 'Document is locked and cannot be deleted');
 		}
-
+		// Perform the delete operation
+		const result = await this.deleteOne({ _id: { $eq: id } }, options);
+		// Check if the delete operation was acknowledged
+		if (!result.acknowledged) throw new HttpException(HttpStatus.INTERNAL_SERVER_ERROR, 'Failed to delete documents', result);
+		// Return the result of the delete operation
 		return result;
 	}
 
@@ -105,11 +107,7 @@ export abstract class MongoCollectionClass<T extends Document, TCreate, TUpdate>
 	 */
 	public async deleteMany(filter: Filter<T>): Promise<DeleteResult> {
 		const result = await this.mongoCollection.deleteMany(filter);
-
-		if (!result.acknowledged) {
-			throw new HttpException(HttpStatus.INTERNAL_SERVER_ERROR, 'Failed to delete documents', result);
-		}
-
+		if (!result.acknowledged) throw new HttpException(HttpStatus.INTERNAL_SERVER_ERROR, 'Failed to delete documents', result);
 		return result;
 	}
 
@@ -118,13 +116,16 @@ export abstract class MongoCollectionClass<T extends Document, TCreate, TUpdate>
 	 * @param filter - The filter criteria to match the document to delete
 	 * @returns A promise that resolves to the result of the delete operation
 	 */
-	public async deleteOne(filter: Filter<T>, options?: DeleteOptions): Promise<DeleteResult> {
-		const result = await this.mongoCollection.deleteOne(filter, options);
-
-		if (!result.acknowledged) {
-			throw new HttpException(HttpStatus.INTERNAL_SERVER_ERROR, 'Failed to delete document', result);
+	public async deleteOne(filter: Filter<T>, options?: DeleteOptions & { forceIfLocked?: boolean }): Promise<DeleteResult> {
+		// If forceIfLocked is not set then check if the document is locked.
+		// If it is locked, then throw an error to prevent the operation.
+		if (!options?.forceIfLocked) {
+			const isLocked = await this.isLocked(filter);
+			if (isLocked) throw new HttpException(HttpStatus.FORBIDDEN, 'Document is locked and cannot be deleted');
 		}
-
+		// Perform the delete operation
+		const result = await this.mongoCollection.deleteOne(filter, options);
+		if (!result.acknowledged) throw new HttpException(HttpStatus.INTERNAL_SERVER_ERROR, 'Failed to delete document', result);
 		return result;
 	}
 
@@ -158,30 +159,29 @@ export abstract class MongoCollectionClass<T extends Document, TCreate, TUpdate>
 
 	/**
 	 * Checks if a document with the given ID exists in the collection.
-	 * @param _id The ID of the document to check for existence.
+	 * @param id The ID of the document to check for existence.
 	 * @returns A promise that resolves to true if the document exists, false otherwise.
 	 */
-	public async existsById(_id: T['_id']): Promise<boolean> {
-		const doc = await this.mongoCollection.findOne({ _id: _id }, { projection: { _id: 1 } });
-		return doc !== null;
+	public async existsById(id: T['_id']): Promise<boolean> {
+		const foundDoc = await this.mongoCollection.findOne({ _id: id }, { projection: { _id: 1 } });
+		return foundDoc !== null;
 	}
 
 	/**
 	 * Finds a document by its ID.
-	 * @param _id The ID of the document to find.
+	 * @param id The ID of the document to find.
 	 * @param options Optional find options.
-	 * @returns A promise that resolves to the matching document or null if not found
+	 * @returns A promise that resolves to the matching document or null if not found.
 	 */
-	public async findById(_id: T['_id'], options?: FindOptions): Promise<null | WithId<T>> {
-		const filter: Filter<T> = { _id: { $eq: _id } } as Filter<T>;
-		return this.mongoCollection.findOne(filter, options);
+	public async findById(id: T['_id'], options?: FindOptions): Promise<null | WithId<T>> {
+		return this.mongoCollection.findOne({ _id: { $eq: id } }, options);
 	}
 
 	/**
 	 * Finds multiple documents matching the filter criteria with optional pagination and sorting.
-	 * @param filter - (Optional) filter criteria to match documents
-	 * @param options - (Optional) find options
-	 * @returns A promise that resolves to an array of matching documents
+	 * @param filter (Optional) filter criteria to match documents.
+	 * @param options (Optional) find options.
+	 * @returns A promise that resolves to an array of matching documents.
 	 */
 	public async findMany(filter?: Filter<T>, options?: FindOptions): Promise<WithId<T>[]> {
 		return await this.mongoCollection.find(filter ?? {}, options).toArray();
@@ -189,9 +189,9 @@ export abstract class MongoCollectionClass<T extends Document, TCreate, TUpdate>
 
 	/**
 	 * Finds a single document matching the filter criteria.
-	 * @param filter - The filter criteria to match the document
-	 * @param options - (Optional) find options
-	 * @returns A promise that resolves to the matching document or null if not found
+	 * @param filter The filter criteria to match the document.
+	 * @param options (Optional) find options.
+	 * @returns A promise that resolves to the matching document or null if not found.
 	 */
 	public async findOne(filter: Filter<T>, options?: FindOptions): Promise<null | WithId<T>> {
 		return await this.mongoCollection.findOne(filter, options);
@@ -315,15 +315,67 @@ export abstract class MongoCollectionClass<T extends Document, TCreate, TUpdate>
 	}
 
 	/**
+	 * Checks if a document with the given ID is locked or not.
+	 * @param id The ID of the document to check.
+	 * @returns A promise that resolves to the result of the check operation.
+	 */
+	public async isLocked(filter: Filter<T>): Promise<boolean> {
+		// Fetch the document by its ID from the database
+		const foundDoc = await this.findOne(filter);
+		// If the document has a is_locked field and it resolves to a truthy value,
+		// then the document is considered locked.
+		if (foundDoc?.is_locked) return true;
+		// Otherwise, the document is not locked.
+		return false;
+	}
+
+	/**
+	 * Checks if a document with the given ID is locked or not.
+	 * @param id The ID of the document to check.
+	 * @returns A promise that resolves to the result of the check operation.
+	 */
+	public async isLockedById(id: T['_id']): Promise<boolean> {
+		// Fetch the document by its ID from the database
+		const foundDoc = await this.findById(id);
+		// If the document has a is_locked field and it resolves to a truthy value,
+		// then the document is considered locked.
+		if (foundDoc?.is_locked) return true;
+		// Otherwise, the document is not locked.
+		return false;
+	}
+
+	/**
+	 * Toogle the lock status of a document by its ID.
+	 * @param id The ID of the document to toggle lock status.
+	 * @param forceValue Optional boolean to explicitly set the lock status.
+	 * @returns A promise that resolves to the result of the update operation.
+	 */
+	public async toggleLockById(id: T['_id'], forceValue?: boolean): Promise<void> {
+		// Get the current document from the database
+		const foundDoc = await this.findById(id);
+		if (!foundDoc) throw new Error('Document not found');
+		// Determine the new lock status
+		const newLockStatus = forceValue !== undefined ? forceValue : !foundDoc.is_locked;
+		// Update the document with the new lock status
+		await this.mongoCollection.updateOne({ _id: { $eq: id } }, { $set: { is_locked: newLockStatus } } as unknown as Partial<T>);
+	}
+
+	/**
 	 * Updates a document by its ID.
-	 * @param _id The ID of the document to update.
+	 * @param id The ID of the document to update.
 	 * @param updateFields The fields to update in the document.
 	 * @param options Optional options for the update operation.
 	 * @returns A promise that resolves to the result of the update operation.
 	 */
-	public async updateById<TReturnDocument extends boolean = true>(_id: T['_id'], updateFields: TUpdate, options?: UpdateOptions & { returnResult?: TReturnDocument }): Promise<TReturnDocument extends true ? WithId<T> : UpdateResult<T>> {
-		const filter: Filter<T> = { _id: { $eq: _id } } as Filter<T>;
-		return this.updateOne(filter, updateFields, options);
+	public async updateById<TReturnDocument extends boolean = true>(id: T['_id'], updateFields: TUpdate, options?: UpdateOptions & { forceIfLocked?: boolean, returnResult?: TReturnDocument }): Promise<TReturnDocument extends true ? WithId<T> : UpdateResult<T>> {
+		// If forceIfLocked is not set then check if the document is locked.
+		// If it is locked, then throw an error to prevent the operation.
+		if (!options?.forceIfLocked) {
+			const isLocked = await this.isLockedById(id);
+			if (isLocked) throw new HttpException(HttpStatus.FORBIDDEN, 'Document is locked and cannot be updated');
+		}
+		// Perform the update operation
+		return this.updateOne({ _id: { $eq: id } }, updateFields, options);
 	}
 
 	/**
@@ -368,12 +420,19 @@ export abstract class MongoCollectionClass<T extends Document, TCreate, TUpdate>
 
 	/**
 	 * Updates a single document matching the filter criteria.
-	 * @param filter - The filter criteria to match the document to update
-	 * @param updateFields - The fields to update in the document
-	 * @param options - The options for the update operation
-	 * @returns A promise that resolves to the result of the update operation
+	 * @param filter The filter criteria to match the document to update.
+	 * @param updateFields The fields to update in the document.
+	 * @param options The options for the update operation.
+	 * @returns A promise that resolves to the result of the update operation.
 	 */
-	public async updateOne<TReturnDocument extends boolean = true>(filter: Filter<T>, updateFields: TUpdate, options?: UpdateOptions & { returnResult?: TReturnDocument }): Promise<TReturnDocument extends true ? WithId<T> : UpdateResult<T>> {
+	public async updateOne<TReturnDocument extends boolean = true>(filter: Filter<T>, updateFields: TUpdate, options?: UpdateOptions & { forceIfLocked?: boolean, returnResult?: TReturnDocument }): Promise<TReturnDocument extends true ? WithId<T> : UpdateResult<T>> {
+		// If forceIfLocked is not set then check if the document is locked.
+		// If it is locked, then throw an error to prevent the operation.
+		if (!options?.forceIfLocked) {
+			const isLocked = await this.isLocked(filter);
+			if (isLocked) throw new HttpException(HttpStatus.FORBIDDEN, 'Document is locked and cannot be updated');
+		}
+		// Perform the update operation
 		let parsedUpdateFields = updateFields;
 		if (this.updateSchema) {
 			try {
